+++ src/msmail/cli.py
@@ -10,6 +10,7 @@ from msmail.commands import move as move_command
 from msmail.commands import read as read_command
 from msmail.commands import save_attachments as save_attachments_command
 from msmail.commands import search as search_command
+from msmail.commands import send as send_command
 from msmail.commands import respond as respond_command
 from msmail.commands import smime as smime_command
 
@@ -30,6 +31,7 @@ app.command("move")(move_command.move_message)
 app.command("mark")(mark_command.mark_message)
 app.command("save-attachments")(save_attachments_command.save_attachments)
 app.command("search")(search_command.search_messages)
+app.command("send")(send_command.send_message)
 app.command("folders")(search_command.list_folders)
 app.command("reply")(respond_command.reply_message)
 app.command("forward")(respond_command.forward_message)